chore: add raise for status for better error messages
